+++ docs_src/middleware/protocols.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from esmerald.utils.concurrency import AsyncExitStack
 from esmerald.core.config import AsyncExitConfig
-from esmerald.protocols.middleware import MiddlewareProtocol
+from esmerald.core.protocols.middleware import MiddlewareProtocol
 from esmerald.types import ASGIApp, Receive, Scope, Send