+++ esmerald/core/config/__init__.py
@@ -1,6 +1,7 @@
 from .asyncexit import AsyncExitConfig
 from .cors import CORSConfig
 from .csrf import CSRFConfig
+from .logging import LoggingConfig
 from .openapi import OpenAPIConfig
 from .session import SessionConfig
 from .static_files import StaticFilesConfig
@@ -12,4 +13,5 @@ __all__ = [
     "OpenAPIConfig",
     "SessionConfig",
     "StaticFilesConfig",
+    "LoggingConfig",
 ]