+++ esmerald/logging.py
@@ -1,22 +1,25 @@
-import logging
+from __future__ import annotations
 
-from loguru import logger as loguru_logger
+from typing import Any, cast
+
+from esmerald.core.protocols.logging import LoggerProtocol
 
 
-class InterceptHandler(logging.Handler):  # pragma: no cover
-    def emit(self, record: logging.LogRecord) -> None:
-        level: str
-        try:
-            level = loguru_logger.level(record.levelname).name
-        except ValueError:
-            level = str(record.levelno)
+class LoggerProxy:
+    """
+    Proxy for the real logger used by Esmerald.
+    """
 
-        frame, depth = logging.currentframe(), 2
-        while frame.f_code.co_filename == logging.__file__:
-            frame = frame.f_back
-            depth += 1
+    def __init__(self) -> None:
+        self._logger: LoggerProtocol | None = None
 
-        loguru_logger.opt(depth=depth, exception=record.exc_info).log(
-            level,
-            record.getMessage(),
-        )
+    def bind_logger(self, logger: LoggerProtocol | None) -> None:  # noqa
+        self._logger = logger
+
+    def __getattr__(self, item: str) -> Any:
+        if not self._logger:
+            raise RuntimeError("Logger is not configured yet. Please call setup_logging() first.")
+        return getattr(self._logger, item)
+
+
+logger: LoggerProtocol = cast(LoggerProtocol, LoggerProxy())